Almost Finished color extraction. Only thing left to do is get it to work with AlbumGrid
